requirements_1.txt
python-telegram-bot==21.10
pandas==2.1.4 
openpyxl==3.1.2 
numpy==1.26.0  
python-dotenv==1.0.0 